change in datatype for load

+++ ServerTrack/ServerTrack.Tests/TestUtilities.cs
@@ -10,14 +10,14 @@ namespace ServerTrack1.Tests
         [TestMethod]
         public void Test_GetAverageSimple()
         {
-            int result = Utilities.GetAverage(100, 5, 40);
+            double result = Utilities.GetAverage(100, 5, 40);
             Assert.AreEqual(90, result);
         }
 
         [TestMethod]
         public void Test_GetAverageLarge()
         {
-            int result = Utilities.GetAverage(10000000, 5, 40);
+            double result = Utilities.GetAverage(10000000, 5, 40);
             Assert.AreEqual(8333340, result);
         }
     }
